+++ core/src/main/scala/sttp/tapir/FileRange.scala
@@ -3,9 +3,6 @@ package sttp.tapir
 import sttp.model.ContentRangeUnits
 import sttp.model.headers.ContentRange
 
-import java.io.InputStream
-import java.net.URL
-
 case class FileRange(file: TapirFile, range: Option[RangeValue] = None)
 
 case class RangeValue(start: Option[Long], end: Option[Long], fileSize: Long) {
@@ -24,19 +21,5 @@ case class RangeValue(start: Option[Long], end: Option[Long], fileSize: Long) {
     case (Some(_start), None)       => Some((_start, fileSize))
     case (None, Some(_end))         => Some((fileSize - _end, fileSize))
     case _                          => None
-  }
-}
-
-case class InputStreamRange(inputStream: () => InputStream, range: Option[RangeValue] = None) {
-  def inputStreamFromRangeStart: () => InputStream = range.flatMap(_.start) match {
-    case Some(start) if start > 0 =>
-      () =>
-        val openedStream = inputStream()
-        val skipped = openedStream.skip(start)
-        if (skipped == start)
-          openedStream
-        else
-          throw new IllegalArgumentException(s"Illegal range start: $start, could skip only $skipped bytes")
-    case _ => inputStream
   }
 }
